integrate 'specify-string' and use for validation; update docs; fix issue with 'printWithOptions'
