+++ package.json
@@ -33,7 +33,8 @@
   "dependencies": {
     "@liquid-labs/condition-eval": "^1.0.0-alpha.18",
     "cli-columns": "^4.0.0",
-    "magic-print": "^1.0.0-alpha.7"
+    "magic-print": "^1.0.0-alpha.18",
+    "specify-string": "^1.0.3"
   },
   "devDependencies": {
     "@liquid-labs/sdlc-resource-babel-and-rollup": "^1.0.0-alpha.5",
