+++ samples/demo.json
@@ -4,9 +4,11 @@
     { 
       "prompt": "What are your 2 or 3 favorite colors?",
       "multiValue": true,
-      "requireMinCount": 2,
-      "requireMaxCount": 3,
-      "parameter": "FAVORITE_COLOR"
+      "parameter": "FAVORITE_COLOR",
+      "validations": {
+        "min-count": 2,
+        "max-count": 3
+      }
     },
     {
       "prompt": "What's your favorite integer number?",
